Feature/Write-Unit-Test-With-Jest

+++ src/modules/products/products.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateProductParams, UpdateProductParams } from './utils/type';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Product } from 'src/entities/Product';
+import { Product } from '../../entities/Product';
 import { Repository } from 'typeorm';
 
 @Injectable()
@@ -15,20 +15,35 @@ export class ProductsService {
     }
 
     async findProduct(id){
-        return await this.productRepository.findOneBy(id);
+        const product =  await this.productRepository.findOneBy(id);
+        if(!product){
+            throw new NotFoundException(`Product with id ${id} not found`);
+        }
+        return product;
     }
 
     async createProduct(createProductDetails: CreateProductParams) {
         const newProduct = this.productRepository.create(createProductDetails);
+        if(!newProduct){
+            throw new BadRequestException();
+        }
         return await this.productRepository.save(newProduct);
         
     }
 
-    async updateProduct(id: number, updateProductDetails : UpdateProductParams) {
+    async updateProduct(id, updateProductDetails : UpdateProductParams) {
+        const product = await this.productRepository.findOneBy(id);
+        if(!product){
+            throw new NotFoundException(`Product with id ${id} not found`);
+        }
         return await this.productRepository.update({id}, updateProductDetails);
     }
 
     async deleteProduct(id) {
+        const product = await this.productRepository.findOneBy(id);
+        if(!product){
+            throw new NotFoundException(`Product with id ${id} not found`);
+        }
         return await this.productRepository.delete(id);
     }
 }
